reposition image location
