+++ resources/player/player_up-3.png
